Move remaining UI classes into appropriate packages
