+++ Source/Android/app/src/main/java/org/dolphinemu/dolphinemu/ui/files/FileViewHolder.java
@@ -1,4 +1,4 @@
-package org.dolphinemu.dolphinemu.viewholders;
+package org.dolphinemu.dolphinemu.ui.files;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
