+++ Source/Android/app/src/main/java/org/dolphinemu/dolphinemu/ui/ingame/LoadStateFragment.java
@@ -1,4 +1,4 @@
-package org.dolphinemu.dolphinemu.fragments;
+package org.dolphinemu.dolphinemu.ui.ingame;
 
 import android.app.Fragment;
 import android.os.Bundle;
@@ -11,7 +11,6 @@ import android.widget.GridLayout;
 
 import org.dolphinemu.dolphinemu.BuildConfig;
 import org.dolphinemu.dolphinemu.R;
-import org.dolphinemu.dolphinemu.activities.EmulationActivity;
 
 public final class LoadStateFragment extends Fragment implements View.OnClickListener
 {
