+++ Source/Android/app/src/main/java/org/dolphinemu/dolphinemu/ui/main/tv/TvSettingsViewHolder.java
@@ -1,4 +1,4 @@
-package org.dolphinemu.dolphinemu.viewholders;
+package org.dolphinemu.dolphinemu.ui.main.tv;
 
 
 import android.support.v17.leanback.widget.ImageCardView;
