+++ Source/Android/app/src/main/res/layout-television/fragment_emulation.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="org.dolphinemu.dolphinemu.fragments.EmulationFragment">
+             tools:context=".ui.ingame.EmulationFragment">
 
     <!-- This is what everything is rendered to during emulation -->
     <SurfaceView
