+++ Source/Android/app/src/main/res/layout/fragment_emulation.xml
@@ -3,7 +3,7 @@
              android:layout_width="match_parent"
              android:layout_height="match_parent"
              android:keepScreenOn="true"
-             tools:context="org.dolphinemu.dolphinemu.fragments.EmulationFragment">
+             tools:context=".ui.ingame.EmulationFragment">
 
     <!-- This is what everything is rendered to during emulation -->
     <SurfaceView
@@ -14,7 +14,7 @@
         android:focusableInTouchMode="false"/>
 
     <!-- This is the onscreen input overlay -->
-    <org.dolphinemu.dolphinemu.overlay.InputOverlay
+    <org.dolphinemu.dolphinemu.ui.ingame.overlay.InputOverlay
         android:id="@+id/surface_input_overlay"
         android:layout_height="match_parent"
         android:layout_width="match_parent"
